+++ app/src/main/java/com/example/meetup/model/PostReviewWriteResponseModel.kt
@@ -4,10 +4,10 @@ data class PostReviewWriteResponseModel(
     var isSuccess: Boolean,
     var code: Long,
     var message: String,
-    var result: result
+    var result: PostReviewWriteResult
 )
 
-data class result(
+data class PostReviewWriteResult(
     var id: Long,
     var orderId: Long,
     var rating: Long,
