+++ app/src/main/java/com/example/meetup/model/review/PostReviewWriteRequestModel.kt
@@ -1,0 +1,15 @@
+package com.example.meetup.model.review
+
+data class PostReviewWriteModel (
+    var dto : PostReviewWriteModelDto,
+
+    var imge : String
+)
+
+public data class PostReviewWriteModelDto (
+
+    var orderId : Long,
+    var rating : Long,
+    var message : String
+)
+
