+++ app/src/main/java/com/example/meetup/model/review/ReviewListResponseModel.kt
@@ -1,4 +1,4 @@
-package com.example.meetup.model
+package com.example.meetup.model.review
 
 data class ReviewListResponseModel(
     var reviewMenuOptionName : String,
